requirements.txt
fastapi
uvicorn
opencv-python
numpy
ultralytics
python-multipart
pydantic
sqlalchemy
aiosqlite
pillow
imageio
imageio-ffmpeg
httpx
websockets
psutil
openai>=1.0.0
pyyaml>=6.0
paho-mqtt>=1.6.1
python-telegram-bot
